Bug fix for negative numbers in seq and ack for tcp header. See #27

+++ src/main/java/com/bits/protocolanalyzer/analyzer/transport/TcpAnalyzer.java
@@ -47,12 +47,12 @@ public class TcpAnalyzer extends TransportAnalyzer {
     }
 
     @Override
-    public int getAckNumber() {
+    public long getAckNumber() {
         return TcpHeader.getAckNumber(this.tcpHeader);
     }
 
     @Override
-    public int getSeqNumber() {
+    public long getSeqNumber() {
         return TcpHeader.getSequenceNumber(this.tcpHeader);
     }
